Make JiraClient methods take project_key arg
Rather than initializing the JiraClient to work with on particular
project key, we now initialize it to be project agnostic. This is a step
towards versioning related projects as a batch, rather than doing one
project at a time. As a result, we now have to pass `project_key` as an
argument to the Jira API wrapper methods that work with specific
projects.