Add some gitclient methods for version cards
gitclient will need some refactoring once more of the client-code
requirements are fleshed out. `project_name` is being repeated in a ton
of places; it would probably just be easier to create our own `Repo`
class to hold the information we need accessible about specific repos.

The other thing is that there's not a clear distinction between local
and remote repo actions. Such a distinction might be helpful, but for
now its not certain.