WIP First pass at making `!sign` work
So far the most noticeable issue seems to be that the commit SHA
included in the version name points to an earlier commit instead of the
merge commit that represents the release.

Other than that, the code is a bit messy and can probably be made much
more DRY, not only on its own, but also WRT the `!send` and `!seal`
commands.