Add methods for reading tags from local repo
The methods are not yet in use; this is a transition commit. The problem
with retrieving tags from origin using pygithub is that GitHub's REST
API is very verbose, sending too much data and taking too long. Git is
designed to be concise, so we may as well depend on it for this
function. The downside of using Git through a subprocess instead of
pygithub is that it forces us to parse git stdout, which may not be as
clean as getting the information from the REST API.

A third alternative, using GitHub's GraphQL API (v4), has not been tested
yet. Part of the reason for that is that API v4 does not yet support all
of the methods in the REST API (v3).