Get release_type earlier
When finalizing a release, a Jira version is created, set on all
eligible tickets, and then renamed to include the commit sha from the
merge commit. Under this flow, the release_type calculation is not
possible once the initial Version is created. This is a workaround which
gets the release_type prior to creating the Jira version. The retrieved
release_type is then passed in as an argument.